complete the ticket
